requirements.txt
flask
flask-jwt-extended
mysql-connector-python
werkzeug
python-dotenv
